move executor tests to it to prevent circular dep issue - move test helper methods to test-utils
